+++ src/BitBayPayClient/IBitBayPayService.cs
@@ -10,7 +10,7 @@ namespace BitBayPayClient
             PaymentType paymentType);
 
         Task<BitBayPayResponse<CreatePaymentResponse>> CreatePayment(string destinationCurrency,
-            double price,
+            decimal price,
             Guid orderId,
             string? sourceCurrency,
             string successCallbackUrl,
